# Use SSE 4.2 due to scan string in read_str()